Added account tab
