+++ front_end/src/app/(main)/accounts/settings/account/components/api_access.tsx
@@ -8,6 +8,7 @@ import React, { FC, useState } from "react";
 import toast from "react-hot-toast";
 
 import Button from "@/components/ui/button";
+import { Input } from "@/components/ui/form_field";
 
 export type Props = {
   token: string;
@@ -17,20 +18,27 @@ const ApiAccess: FC<Props> = ({ token }) => {
   const t = useTranslations();
   const [visible, setVisible] = useState(false);
 
+  // TODO: ensure all articles/pages which might have references to the API key page have updated urls
+
   return (
-    <section className="text-sm">
-      <hr />
-      <h2 className="mb-5 mt-3 px-1"> {t("apiAccess")}</h2>
+    <section>
+      <hr className="my-6 border-gray-400 dark:border-gray-400-dark" />
+      <div className="mb-4 text-gray-500 dark:text-gray-500-dark">
+        {t("apiAccess")}
+      </div>
       <div className="text-sm">
-        <p>
+        <p className="my-4">
           {t("apiAcessText")} <Link href="/api">{t("documentation")}</Link>
         </p>
-        <span>{t("yourAPITokenIs")}</span>
-        <div className="mt-2 flex items-center gap-8">
-          <div>
-            <pre>{visible ? token : "*".repeat(40)}</pre>
-          </div>
-          <div className="flex gap-1">
+        <div className="mb-3">{t("yourAPITokenIs")}</div>
+        <div className="flex items-center gap-2.5">
+          <Input
+            className="dark:disabled-text-gray-600-dark block w-full max-w-72 rounded border border-gray-700 bg-inherit p-2.5 disabled:border-gray-400 disabled:bg-gray-200 disabled:text-gray-600 dark:border-gray-700-dark disabled:dark:border-gray-400-dark disabled:dark:bg-gray-200-dark"
+            value={token}
+            disabled={!visible}
+            type={visible ? "text" : "password"}
+          />
+          <div className="flex gap-2.5">
             <Button
               aria-label={t("showApiToken")}
               variant="tertiary"
