+++ front_end/src/app/(main)/accounts/settings/notifications/components/email_notifications.tsx
@@ -6,8 +6,7 @@ import { useTranslations } from "next-intl";
 import React, { FC, useCallback, useState } from "react";
 
 import { updateProfileAction } from "@/app/(main)/accounts/profile/actions";
-import ChangeEmailModal from "@/app/(main)/accounts/settings/components/change_email";
-import Button from "@/components/ui/button";
+import PreferencesSection from "@/app/(main)/accounts/settings/components/preferences_section";
 import Checkbox from "@/components/ui/checkbox";
 import LoadingSpinner from "@/components/ui/loading_spiner";
 import Tooltip from "@/components/ui/tooltip";
@@ -21,9 +20,7 @@ export type Props = {
 
 const EmailNotifications: FC<Props> = ({ user }) => {
   const t = useTranslations();
-
   const [loadingIndex, setLoadingIndex] = useState<number | null>(null);
-  const [isChangeEmailModalOpen, setIsChangeEmailModalOpen] = useState(false);
 
   const handleEmailSubscriptionChange = useCallback(
     async (subscriptionType: SubscriptionEmailType, checked: boolean) => {
@@ -87,41 +84,28 @@ const EmailNotifications: FC<Props> = ({ user }) => {
   ];
 
   return (
-    <section className="text-sm">
-      <hr />
-      <h2 className="mb-5 mt-3 px-1">{t("settingsSubscriptions")}</h2>
-      <h3 className="bg-blue-200 p-1 text-sm font-medium dark:bg-blue-200-dark">
-        {t("settingsUserEmail")}
-      </h3>
-      <div className="p-1 text-sm">
-        <span className="pr-8">{user.email}</span>{" "}
-        <Button variant="link" onClick={() => setIsChangeEmailModalOpen(true)}>
-          {t("edit")}
-        </Button>
-        <ChangeEmailModal
-          isOpen={isChangeEmailModalOpen}
-          onClose={() => setIsChangeEmailModalOpen(false)}
-        />
+    <PreferencesSection title={t("settingsEmailNotifications")}>
+      <div className="flex flex-col gap-3">
+        {options.map(({ type, ...opts }, index) => (
+          <div className="flex items-center" key={`subscriptions-${type}`}>
+            <Checkbox
+              checked={!user.unsubscribed_mailing_tags.includes(type)}
+              onChange={(checked) => {
+                updateProfile(type, checked);
+              }}
+              onClick={() => setLoadingIndex(index)}
+              className="p-1"
+              readOnly={isPending}
+              inputClassName="text-gray-900 dark:text-gray-900-dark"
+              {...opts}
+            />
+            {loadingIndex === index && isPending && (
+              <LoadingSpinner size="1x" />
+            )}
+          </div>
+        ))}
       </div>
-      <h3 className="bg-blue-200 p-1 text-sm font-medium dark:bg-blue-200-dark">
-        {t("settingsEmailNotifications")}
-      </h3>
-      {options.map(({ type, ...opts }, index) => (
-        <div className="flex items-center" key={`subscriptions-${type}`}>
-          <Checkbox
-            checked={!user.unsubscribed_mailing_tags.includes(type)}
-            onChange={(checked) => {
-              updateProfile(type, checked);
-            }}
-            onClick={() => setLoadingIndex(index)}
-            className="p-1.5"
-            readOnly={isPending}
-            {...opts}
-          />
-          {loadingIndex === index && isPending && <LoadingSpinner size="1x" />}
-        </div>
-      ))}
-    </section>
+    </PreferencesSection>
   );
 };
