--- front_end/src/app/(main)/accounts/settings2/account/page.tsx
@@ -1,7 +1,0 @@
-export default async function Page() {
-  return (
-    <>
-      <h2>Account Settings</h2>
-    </>
-  );
-}
